Refactored existing files code, added birthdays feature
- Refactored code in existing files for better readability and performance
- Added "birthdays" feature:
  - Users can now save birthdays
  - Application sends reminders for saved birthdays
- Implemented listing of 10 upcoming birthdays:
  - Displays a list of the next 10 upcoming birthdays
  - Sorted by date for easy viewing

+++ src/commands/admin/autorole-disable.js
@@ -1,4 +1,8 @@
+const { EmbedBuilder } = require('discord.js');
 const AutoRole = require('../../models/AutoRole');
+
+const errorEmbed = new EmbedBuilder()
+  .setColor('#FF0000');
 
 module.exports = {
   data: {
@@ -7,20 +11,27 @@ module.exports = {
   },
 
   run: async ({ interaction }) => {
+    if (!interaction.inGuild()) {
+      errorEmbed.setDescription('You can only run this command inside a server.');
+      await interaction.reply({ embeds: [errorEmbed] });
+      return;
+    }
+
     try {
       await interaction.deferReply();
 
       if (!(await AutoRole.exists({ guildId: interaction.guild.id }))) {
-        interaction.editReply(
-          'Autorole nie są skonfigurowane. Aby skonfigurować, uruchom `/autorole-configure`.'
-        );
+        errorEmbed.setDescription('Autorole nie są skonfigurowane. Aby skonfigurować, uruchom `/autorole-configure`.');
+        await interaction.reply({ embeds: [errorEmbed] });
         return;
       }
 
       await AutoRole.findOneAndDelete({ guildId: interaction.guild.id });
-      interaction.editReply(
-        'Autorole zostały wyłączone dla tego serwera. Aby skonfigurować, uruchom `/autorole-configure`.'
-      );
+
+      const successEmbed = new EmbedBuilder()
+        .setColor('#00BFFF')
+        .setDescription('Autorole zostały wyłączone dla tego serwera. Aby skonfigurować, uruchom `/autorole-configure`.');
+      await interaction.reply({ embeds: [successEmbed] });
     } catch (error) {
       console.log(`Wystąpił błąd podczas wyłączania autoroli dla tego serwera: ${error}`);
     }
